Send a notification when the door is closed
If the door is closed after being open for longer than the notification
time (presumably triggering an alert) then another notification is sent
when it is closed. This helps make it clear that the door is closed
again if you have left after say packing the car for a while.